Feature/bug 25 (#27)
* change augmentations and add image clipping

* change image to uint8

* fixes #25

+++ ultrayolo/datasets/datasets.py
@@ -342,7 +342,7 @@ class CocoFormatDataset(tf.keras.utils.Sequence):
                 img_path = self.images_path / doc['file_name']
                 img = common.open_image(img_path)
                 assert np.all(img >= 0)
-                batch_images.append(img)
+                batch_images.append(img.astype(np.uint8))
 
                 boxes = []
                 classes = []
@@ -368,6 +368,6 @@ class CocoFormatDataset(tf.keras.utils.Sequence):
                                                   self.num_classes,
                                                   self.target_shape, classes)
 
-            return batch_images.astype(np.float32), batch_boxes
+            return batch_images, batch_boxes
 
         return batch_images, batch_boxes, batch_classes
